Compares the benchmarks for a PR to the latest master benchmarks (example below). 

Based on the "results comparer" tool in the [dotnet/performance repo](https://github.com/dotnet/performance/blob/main/src/tools/ResultsComparer/README.md), using a default threshold (5%) and noise levels (0.3ns). We may well need to tweak these for our cases - we'll have to keep an eye on that.

@DataDog/apm-dotnet


Example: 

---

## Benchmarks Report :snail:

Benchmarks for #123 compared to [master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b): 
* 2 benchmarks are **better**, with geometric mean 1.118
* 3 benchmarks are **worse**, with geometric mean 1.136

### Benchmark details


<details><summary><strong>Benchmarks.Trace.AgentWriterBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`WriteAndFlushEnrichedTraces`|net472|952.0&nbsp;μs|17.42&nbsp;μs|16.29&nbsp;μs|0.0000| 0.0000|0.0000|3.09&nbsp;KB|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`WriteAndFlushEnrichedTraces`|netcoreapp3.1|736.9&nbsp;μs|10.90&nbsp;μs|9.66&nbsp;μs|0.0000| 0.0000|0.0000|2.51&nbsp;KB|
|#123|`WriteAndFlushEnrichedTraces`|net472|981.4&nbsp;μs|18.93&nbsp;μs|25.91&nbsp;μs|0.00| 0.0000|0.0000|0.0000|
|#123|`WriteAndFlushEnrichedTraces`|netcoreapp3.1|769.7&nbsp;μs|14.90&nbsp;μs|20.39&nbsp;μs|0.03| 0.0000|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.AspNetCoreBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`SendRequest`|net472|0.0000&nbsp;ns|0.0000&nbsp;ns|0.0000&nbsp;ns|?| 0.0000|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`SendRequest`|netcoreapp3.1|296,099.8976&nbsp;ns|5,361.8618&nbsp;ns|6,584.8478&nbsp;ns|?| 0.1443|0.0000|0.0000|
|#123|`SendRequest`|net472|0.0000&nbsp;ns|0.0000&nbsp;ns|0.0000&nbsp;ns|?| 0.0000|0.0000|0.0000|
|#123|`SendRequest`|netcoreapp3.1|292,110.1385&nbsp;ns|3,302.7363&nbsp;ns|2,927.7897&nbsp;ns|?| 0.1485|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.DbCommandBenchmark</strong> - Slower :warning:



#### Slower :warning: in #123

|  Benchmark         | diff/base | Base Median (ns) | Diff Median (ns) | Modality   |
|:----------|-----------:|-----------:|--------:|--------:|
| Benchmarks.Trace.DbCommandBenchmark.ExecuteNonQuery&#8209;net472 | 1.076 | 2,063.89 | 2,220.88 | 



#### Faster :tada: in #123

|  Benchmark         | base/diff | Base Median (ns) | Diff Median (ns) | Modality   |
|:----------|-----------:|-----------:|--------:|--------:|
| Benchmarks.Trace.DbCommandBenchmark.ExecuteNonQuery&#8209;netcoreapp3.1 | 1.080 | 2,142.88 | 1,983.23 | 

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`ExecuteNonQuery`|net472|2.071&nbsp;μs|0.0395&nbsp;μs|0.0440&nbsp;μs|0.00| 0.1084|0.0010|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`ExecuteNonQuery`|netcoreapp3.1|2.147&nbsp;μs|0.0414&nbsp;μs|0.0509&nbsp;μs|0.03| 0.0094|0.0000|0.0000|
|#123|`ExecuteNonQuery`|net472|2.223&nbsp;μs|0.0434&nbsp;μs|0.0385&nbsp;μs|0.00| 0.1084|0.0011|0.0000|
|#123|`ExecuteNonQuery`|netcoreapp3.1|1.987&nbsp;μs|0.0395&nbsp;μs|0.0604&nbsp;μs|0.03| 0.0098|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.ElasticsearchBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`CallElasticsearch`|net472|3.293&nbsp;μs|0.0591&nbsp;μs|0.0971&nbsp;μs|0.00| 0.1278|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`CallElasticsearch`|netcoreapp3.1|1.618&nbsp;μs|0.0259&nbsp;μs|0.0230&nbsp;μs|0.02| 0.0104|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`CallElasticsearchAsync`|net472|3.454&nbsp;μs|0.0683&nbsp;μs|0.1159&nbsp;μs|0.00| 0.1502|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`CallElasticsearchAsync`|netcoreapp3.1|1.783&nbsp;μs|0.0345&nbsp;μs|0.0323&nbsp;μs|0.02| 0.0122|0.0000|0.0000|
|#123|`CallElasticsearch`|net472|3.122&nbsp;μs|0.0622&nbsp;μs|0.0831&nbsp;μs|0.00| 0.1280|0.0000|0.0000|
|#123|`CallElasticsearch`|netcoreapp3.1|1.658&nbsp;μs|0.0284&nbsp;μs|0.0265&nbsp;μs|0.02| 0.0108|0.0000|0.0000|
|#123|`CallElasticsearchAsync`|net472|3.232&nbsp;μs|0.0456&nbsp;μs|0.0426&nbsp;μs|0.00| 0.1502|0.0000|0.0000|
|#123|`CallElasticsearchAsync`|netcoreapp3.1|1.745&nbsp;μs|0.0347&nbsp;μs|0.0340&nbsp;μs|0.01| 0.0123|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.GraphQLBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`ExecuteAsync`|net472|3.492&nbsp;μs|0.0696&nbsp;μs|0.0953&nbsp;μs|0.1688| 0.0000|0.0000|1083&nbsp;B|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`ExecuteAsync`|netcoreapp3.1|1.961&nbsp;μs|0.0372&nbsp;μs|0.0311&nbsp;μs|0.0145| 0.0000|0.0000|1008&nbsp;B|
|#123|`ExecuteAsync`|net472|3.452&nbsp;μs|0.0592&nbsp;μs|0.0553&nbsp;μs|0.1685| 0.0000|0.0000|1083&nbsp;B|
|#123|`ExecuteAsync`|netcoreapp3.1|1.851&nbsp;μs|0.0244&nbsp;μs|0.0228&nbsp;μs|0.0142| 0.0000|0.0000|1008&nbsp;B|
</details>


<details><summary><strong>Benchmarks.Trace.HttpClientBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`SendAsync`|net472|5.987&nbsp;μs|0.1127&nbsp;μs|0.1054&nbsp;μs|0.00| 0.3488|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`SendAsync`|netcoreapp3.1|4.333&nbsp;μs|0.0843&nbsp;μs|0.0747&nbsp;μs|0.02| 0.0303|0.0000|0.0000|
|#123|`SendAsync`|net472|6.051&nbsp;μs|0.1137&nbsp;μs|0.1397&nbsp;μs|0.00| 0.3507|0.0000|0.0000|
|#123|`SendAsync`|netcoreapp3.1|4.448&nbsp;μs|0.0816&nbsp;μs|0.0802&nbsp;μs|0.02| 0.0310|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.ILoggerBenchmark</strong> - Slower :warning:



#### Slower :warning: in #123

|  Benchmark         | diff/base | Base Median (ns) | Diff Median (ns) | Modality   |
|:----------|-----------:|-----------:|--------:|--------:|
| Benchmarks.Trace.ILoggerBenchmark.EnrichedLog&#8209;net472 | 1.170 | 3,981.54 | 4,659.79 | 
| Benchmarks.Trace.ILoggerBenchmark.EnrichedLog&#8209;netcoreapp3.1 | 1.163 | 3,867.33 | 4,496.93 | 

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|net472|3.992&nbsp;μs|0.0776&nbsp;μs|0.0924&nbsp;μs|0.00| 0.2872|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|netcoreapp3.1|3.875&nbsp;μs|0.0636&nbsp;μs|0.0564&nbsp;μs|0.03| 0.0279|0.0000|0.0000|
|#123|`EnrichedLog`|net472|4.650&nbsp;μs|0.0909&nbsp;μs|0.0851&nbsp;μs|0.00| 0.2968|0.0000|0.0000|
|#123|`EnrichedLog`|netcoreapp3.1|4.510&nbsp;μs|0.0890&nbsp;μs|0.1025&nbsp;μs|0.04| 0.0290|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.Log4netBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|net472|180.5&nbsp;μs|3.51&nbsp;μs|3.61&nbsp;μs|0.00| 1.2811|0.2745|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|netcoreapp3.1|156.9&nbsp;μs|2.98&nbsp;μs|3.06&nbsp;μs|0.02| 0.0766|0.0000|0.0000|
|#123|`EnrichedLog`|net472|180.3&nbsp;μs|3.46&nbsp;μs|3.70&nbsp;μs|0.00| 1.3094|0.3492|0.0000|
|#123|`EnrichedLog`|netcoreapp3.1|155.1&nbsp;μs|2.76&nbsp;μs|2.58&nbsp;μs|0.03| 0.0755|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.NLogBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|net472|11.76&nbsp;μs|0.233&nbsp;μs|0.311&nbsp;μs|0.00| 0.8541|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|netcoreapp3.1|10.54&nbsp;μs|0.202&nbsp;μs|0.241&nbsp;μs|0.04| 0.0877|0.0000|0.0000|
|#123|`EnrichedLog`|net472|11.91&nbsp;μs|0.238&nbsp;μs|0.326&nbsp;μs|0.00| 0.8535|0.0000|0.0000|
|#123|`EnrichedLog`|netcoreapp3.1|10.30&nbsp;μs|0.201&nbsp;μs|0.198&nbsp;μs|0.03| 0.0913|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.RedisBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`SendReceive`|net472|2.184&nbsp;μs|0.0427&nbsp;μs|0.0333&nbsp;μs|0.00| 0.1547|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`SendReceive`|netcoreapp3.1|1.998&nbsp;μs|0.0388&nbsp;μs|0.0476&nbsp;μs|0.03| 0.0137|0.0000|0.0000|
|#123|`SendReceive`|net472|2.213&nbsp;μs|0.0335&nbsp;μs|0.0313&nbsp;μs|0.00| 0.1543|0.0000|0.0000|
|#123|`SendReceive`|netcoreapp3.1|1.919&nbsp;μs|0.0337&nbsp;μs|0.0282&nbsp;μs|0.02| 0.0137|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.SerilogBenchmark</strong> - Faster :tada:



#### Faster :tada: in #123

|  Benchmark         | base/diff | Base Median (ns) | Diff Median (ns) | Modality   |
|:----------|-----------:|-----------:|--------:|--------:|
| Benchmarks.Trace.SerilogBenchmark.EnrichedLog&#8209;netcoreapp3.1 | 1.156 | 8,319.42 | 7,197.21 | 

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|net472|8.054&nbsp;μs|0.1609&nbsp;μs|0.1506&nbsp;μs|0.00| 0.4480|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`EnrichedLog`|netcoreapp3.1|8.336&nbsp;μs|0.1667&nbsp;μs|0.1853&nbsp;μs|0.03| 0.0370|0.0000|0.0000|
|#123|`EnrichedLog`|net472|7.927&nbsp;μs|0.1383&nbsp;μs|0.1294&nbsp;μs|0.00| 0.4474|0.0000|0.0000|
|#123|`EnrichedLog`|netcoreapp3.1|7.225&nbsp;μs|0.1366&nbsp;μs|0.1277&nbsp;μs|0.02| 0.0377|0.0000|0.0000|
</details>


<details><summary><strong>Benchmarks.Trace.SpanBenchmark</strong> - No change :heavy_check_mark:

</summary>

### Raw results


|  Branch |                      Method |     Toolchain |     Mean |    Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------|---------------------------- |-------------- |---------:|---------:|--------:|-------:|------:|------:|----------:|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`StartFinishSpan`|net472|892.9&nbsp;ns|12.70&nbsp;ns|11.88&nbsp;ns|0.00| 0.0679|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`StartFinishSpan`|netcoreapp3.1|902.9&nbsp;ns|13.39&nbsp;ns|12.52&nbsp;ns|0.02| 0.0058|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`StartFinishScope`|net472|1,083.6&nbsp;ns|19.42&nbsp;ns|18.17&nbsp;ns|0.00| 0.0804|0.0000|0.0000|
|[master](https://github.com/DataDog/dd-trace-dotnet/tree/a6f03eeec8bc8d1313ba50f87a4dd22914a59f0b)|`StartFinishScope`|netcoreapp3.1|1,035.9&nbsp;ns|19.29&nbsp;ns|17.10&nbsp;ns|0.02| 0.0078|0.0000|0.0000|
|#123|`StartFinishSpan`|net472|900.9&nbsp;ns|14.16&nbsp;ns|13.25&nbsp;ns|0.00| 0.0681|0.0000|0.0000|
|#123|`StartFinishSpan`|netcoreapp3.1|857.7&nbsp;ns|12.84&nbsp;ns|12.01&nbsp;ns|0.02| 0.0059|0.0000|0.0000|
|#123|`StartFinishScope`|net472|1,092.3&nbsp;ns|20.42&nbsp;ns|18.10&nbsp;ns|0.00| 0.0806|0.0000|0.0000|
|#123|`StartFinishScope`|netcoreapp3.1|1,023.5&nbsp;ns|20.31&nbsp;ns|23.39&nbsp;ns|0.03| 0.0075|0.0000|0.0000|
</details>